fix(images+api): unify producer→me endpoints & redirects
- API: Moved /api/producer/products → /api/me/products (canonical)
- API: Created thin re-export wrappers at /api/producer/products for compat
- UI: Moved /producer/products pages → /my/products (canonical)
- UI: Created redirect stubs at /producer/products → /my/products
- Updated all API fetch calls to use /api/me/products
- Build: ✅ Passed (46 pages)

Pass 110.1 — Product Image UX Finisher
